requirements-test.txt
-r requirements.txt
pytest==3.3.2
coverage==4.4.2
flake8==3.5.0
pep8-naming==0.5.0
pytest-cov==2.5.1
pytest-mock==1.7.1
vcrpy==1.10
pytest-env==0.6.2
